+++ handlers.go
@@ -58,7 +58,7 @@ func (a *Authenticator) routes(mux *http.ServeMux) {
 			mux.HandleFunc("POST /auth/2fa/webauthn/finish", a.wrap(a.TwoFactorWebauthnFinish))
 		}
 	}
-	if a.socialConfigured("google") || a.socialConfigured("github") || a.socialConfigured("facebook") || a.socialConfigured("apple") {
+	if a.anySocialConfigured() {
 		mux.HandleFunc("GET /auth/social/{provider}/login", a.wrap(a.SocialLogin))
 		mux.HandleFunc("GET /auth/social/{provider}/callback", a.wrap(a.SocialCallback))
 		mux.HandleFunc("POST /auth/social/{provider}/callback", a.wrap(a.SocialCallback)) // Apple posts its callback (form_post)